Overhauled LocalizationDebugWindow UI and added ability to update source
