+++ src/PixiEditor/Data/Localization/LocalizationData.json
@@ -6,7 +6,7 @@
       "code": "en",
       "localeFileName": "en.json",
       "iconFileName": "en.png",
-      "lastUpdated": "2023-05-07 13:31:45"
+      "lastUpdated": "2023-05-08 19:30:00"
     },
     {
       "name": "Polski",
